hv: acpi: remove weak parse_madt
Since it's discarded.

Tracked-On: #861
